[ELY-2034] Refactor utility classes related to OpenID Connect
