+++ http/oidc/src/test/java/org/wildfly/security/http/oidc/OidcTest.java
@@ -53,7 +53,7 @@ import org.wildfly.security.http.HttpServerAuthenticationMechanism;
 import org.wildfly.security.http.HttpServerAuthenticationMechanismFactory;
 import org.wildfly.security.http.HttpServerCookie;
 import org.wildfly.security.http.impl.AbstractBaseHttpTest;
-import org.wildfly.security.json.util.JsonSerialization;
+import org.wildfly.security.jose.util.JsonSerialization;
 
 import com.gargoylesoftware.htmlunit.SilentCssErrorHandler;
 import com.gargoylesoftware.htmlunit.TextPage;
