+++ jose/jwk/pom.xml
@@ -25,6 +25,7 @@
         <groupId>org.wildfly.security</groupId>
         <artifactId>wildfly-elytron-parent</artifactId>
         <version>1.17.0.CR1-SNAPSHOT</version>
+        <relativePath>../../pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
@@ -37,7 +38,7 @@
     <dependencies>
         <dependency>
             <groupId>org.wildfly.security</groupId>
-            <artifactId>wildfly-elytron-json-util</artifactId>
+            <artifactId>wildfly-elytron-jose-util</artifactId>
         </dependency>
         <dependency>
             <groupId>org.wildfly.security</groupId>
