--- json-util/src/main/java/org/wildfly/security/json/util/StringOrArrayDeserializer.java
@@ -1,52 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2021 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.wildfly.security.json.util;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Iterator;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonNode;
-
-/**
- * JSON deserializer for strings or arrays.
- *
- * @author <a href="fjuma@redhat.com">Farah Juma</a>
- */
-public class StringOrArrayDeserializer extends JsonDeserializer<Object> {
-
-    @Override
-    public Object deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
-        JsonNode jsonNode = jsonParser.readValueAsTree();
-        if (jsonNode.isArray()) {
-            ArrayList<String> a = new ArrayList<>(1);
-            Iterator<JsonNode> itr = jsonNode.iterator();
-            while (itr.hasNext()) {
-                a.add(itr.next().textValue());
-            }
-            return a.toArray(new String[a.size()]);
-        } else {
-            return new String[] { jsonNode.textValue() };
-        }
-    }
-
-}
