+++ wildfly-elytron/pom.xml
@@ -361,6 +361,10 @@
             <groupId>org.wildfly.security</groupId>
             <artifactId>wildfly-elytron-jose-jwk</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wildfly.security</groupId>
+            <artifactId>wildfly-elytron-jose-util</artifactId>
+        </dependency>
         <!-- Do not include JWT as it defines a CDI Extension -->
         <!--<dependency>
             <groupId>org.wildfly.security</groupId>
@@ -711,6 +715,11 @@
                                 <dependency>
                                     <groupId>org.wildfly.security</groupId>
                                     <artifactId>wildfly-elytron-jose-jwk</artifactId>
+                                    <version>${project.version}</version>
+                                </dependency>
+                                <dependency>
+                                    <groupId>org.wildfly.security</groupId>
+                                    <artifactId>wildfly-elytron-jose-util</artifactId>
                                     <version>${project.version}</version>
                                 </dependency>
                                 <dependency>
